requirements.txt
APScheduler==3.3.1
Datetime==4.2
Flask==0.12
Flask-Cors==3.0.2
gcloud==0.17.0
gunicorn==19.7.1
google-api-python-client==1.6.2
googleapis-common-protos==1.5.2
line-bot-sdk==1.3.0
oauth2client==3.0.0
protobuf==3.3.0
pycrypto==2.6.1
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
tzlocal==1.4
Werkzeug==0.11.15
wolframalpha==3.0
zope.interface==4.4.2
